HEAT changes, wording, clientside consistency
Reverted ballistics to use table.Copy when generating projectiles.
Normally it isn't an issue, but HEAT modifies it's bulletdata when it
detonates ( Which is stored in ammo crates, causing the next projectile
to inherit the stats of the HEAT slug ).
Changed HEAT fuze var to be compatible with new fuze code.
Changed base HEAT ricochet to 75 from 60.
Changed clientside ballistics to use same method of calculating step
size.